Rename project to sdformat_mjcf
