+++ sdformat_mjcf/sdformat_mjcf/__init__.py
@@ -1,4 +1,4 @@
-from gz_mujoco.gz_mujoco import execute
+from sdformat_mjcf.sdformat_mjcf import execute
 
 
 def main():
